requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
spotipy==2.23.0
pandas==2.1.3
beautifulsoup4==4.12.2
selenium==4.15.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
openai==1.3.5